added TimeStamp Tables, as weel as som other chagnes,

+++ ResturantBusserAPI/Models/Master.cs
@@ -15,7 +15,7 @@ namespace ResturantBusserAPI.Models
         public DateTime RegisterDate { get; set; }
         public String Resturant { get; set; }
         public String Email { get; set; }
-        public  int Phone { get; set; }
+        public  Decimal Phone { get; set; }
         public String Contact { get; set; }
         public String OrgNr { get; set; }
         public String AppId { get; set; }
